Patch of "repelem" function for dense and sparse
The parameters of "repelem" is the same as "repmat", and its behavior is the same as "repelem" function in MATLAB. The idea is from my matlab code, where I frequently use this function.
the files "op_repelem_bones.hpp" and “op_repelem_meat.hpp” is new, and in other files only some related lines are added. What's more, "/include/armadillo" should also be modificated.